letsconnect/chat/api/serializers.py
from rest_framework import serializers

from chat.models import Chat, Profile


class ChatSerializer(serializers.ModelSerializer):
    class Meta:
        model = Chat
        fields = ('__all__')


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = ('__all__')
